remove some unnecessary styles and move files to better directly location
