+++ src/app/components/drawer/Drawer.module.css
@@ -1,12 +1,15 @@
 .drawer {
   --drawerWidth: 256px;
-
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
   transition: width 225ms cubic-bezier(0, 0, .2, 1) 0ms;
   width: var(--drawerWidth);
   z-index: 1;
 
   :global(.MuiDrawer-paperAnchorLeft) {
     border: 0;
+    overflow: hidden;
     padding: 0;
     position: relative;
     width: var(--drawerWidth);
