+++ src/app/components/drawer/DrawerNavigationComponent.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import DrawerNavigationComponent from './DrawerNavigationComponent';
-import DrawerProjects from './drawer/Projects';
+import DrawerProjects from './Projects';
 
 import { mountWithIntl, getStore } from '../../../test/unit/helpers/intl-test';
